+++ libs/vavoomc/vc_expr_invoke.cpp
@@ -2463,9 +2463,9 @@ VExpression *VInvocation::OptimizeBuiltin (VEmitContext &ec) {
       if (!CheckSimpleConstArgs(1, (const int []){TYPE_Vector})) return this;
       v0 = ((VVectorExpr *)Args[0])->GetConstValue();
       if (v0.isValid()) {
-        v0.x = fabs(v0.x);
-        v0.y = fabs(v0.y);
-        v0.z = fabs(v0.z);
+        v0.x = fabsf(v0.x);
+        v0.y = fabsf(v0.y);
+        v0.z = fabsf(v0.z);
         e = new VVectorExpr(v0, Loc);
       } else {
         return this;
